Rename "e2e_simulators" module to "simulators".
