+++ pastis/matrix_generation/matrix_building_numerical.py
@@ -27,9 +27,9 @@ import hcipy
 
 from pastis.config import CONFIG_PASTIS
 import pastis.util as util
-from pastis.e2e_simulators.hicat_imaging import set_up_hicat
-from pastis.e2e_simulators.luvoir_imaging import LuvoirAPLC
-import pastis.e2e_simulators.webbpsf_imaging as webbpsf_imaging
+from pastis.simulators.hicat_imaging import set_up_hicat
+from pastis.simulators.luvoir_imaging import LuvoirAPLC
+import pastis.simulators.webbpsf_imaging as webbpsf_imaging
 import pastis.plotting as ppl
 
 log = logging.getLogger()
